chore: 각 컴포넌트 내 type들을 types 폴더로 이동
