--- src/api/types/auth/index.ts
@@ -1,12 +1,0 @@
-export interface SigninParam {
-  email: string;
-  password: string;
-}
-
-export interface SignupParam {
-  email: string;
-  password: string;
-}
-export interface SigninResponse {
-  access_token: string;
-}
