--- src/api/types/todo/index.ts
@@ -1,5 +1,0 @@
-export interface ModifyTodoParam {
-  isCompleted: boolean;
-  todo: string;
-  todoId: number;
-};
