+++ src/components/common/button/Button.tsx
@@ -1,14 +1,5 @@
-import React, { ButtonHTMLAttributes } from "react";
+import { ButtonProps } from "../../../types/button";
 import StyledButton from "./button.style";
-
-interface CommonComponentProps {
-  children?: React.ReactNode;
-  customStyle?: React.CSSProperties;
-}
-
-export interface ButtonProps
-  extends ButtonHTMLAttributes<HTMLButtonElement>,
-    CommonComponentProps {}
 
 const Button = ({ children, ...atr }: ButtonProps) => {
   return <StyledButton {...atr}>{children}</StyledButton>;
