+++ src/components/common/input/Input.tsx
@@ -1,15 +1,6 @@
 import React from "react";
-import { InputHTMLAttributes } from "react";
+import { InputProps } from "../../../types/input";
 import { StyledInput } from "./input.style";
-
-interface CommonComponentProps {
-  children?: React.ReactNode;
-  customStyle?: React.CSSProperties;
-}
-
-export interface InputProps
-  extends InputHTMLAttributes<HTMLInputElement>,
-    Omit<CommonComponentProps, "children"> {}
 
 const Input = React.forwardRef(
   ({ ...attr }: InputProps, ref: React.ForwardedRef<HTMLInputElement>) => {
